+++ grouping_flutter/lib/app/presentation/components/data_display/key_value_pair_widget.dart
@@ -78,6 +78,8 @@ class KeyValuePairWidget<K, V> extends StatelessWidget{
                 getKeyWidget(context),
                 Gap(gap),
                 getValueWidget(context),
+                if (child != null) Gap(gap),
+                if (child != null) child!,
               ]
             ),
           ),
