+++ grouping_flutter/lib/app/presentation/views/routes.dart
@@ -2,8 +2,8 @@ import 'package:flutter/material.dart';
 import 'package:go_router/go_router.dart';
 import 'package:grouping_project/app/presentation/providers/token_manager.dart';
 import 'package:grouping_project/auth/presentation/views/auth_view.dart';
-import 'package:grouping_project/space/presentation/view_models/user_page_view_model.dart';
-import 'package:grouping_project/space/presentation/view_models/workspace_view_model.dart';
+import 'package:grouping_project/space/presentation/view_models/user_data_provider.dart';
+import 'package:grouping_project/space/presentation/view_models/group_data_provider.dart';
 import 'package:grouping_project/space/presentation/views/pages/user_page_view.dart';
 import 'package:grouping_project/space/presentation/views/pages/workspace_page_view.dart';
 import 'package:provider/provider.dart';
