+++ grouping_flutter/lib/space/presentation/view_models/join_workspace_view_model.dart
@@ -10,7 +10,7 @@ import 'package:grouping_project/space/data/repositories/workspace_repository_im
 import 'package:grouping_project/space/domain/entities/workspace_entity.dart';
 import 'package:grouping_project/space/domain/usecases/workspace_usecases/get_current_workspace_usecase.dart';
 import 'package:grouping_project/space/domain/usecases/workspace_usecases/join_worksapcee_usecase.dart';
-import 'package:grouping_project/space/presentation/view_models/user_page_view_model.dart';
+import 'package:grouping_project/space/presentation/view_models/user_data_provider.dart';
 import 'package:image_picker/image_picker.dart';
 
 class JoinWorkspaceViewModel extends ChangeNotifier {
