+++ grouping_flutter/lib/space/presentation/views/frames/activity_list_frame.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:grouping_project/space/presentation/view_models/activity_list_view_model.dart';
-import 'package:grouping_project/space/presentation/view_models/user_page_view_model.dart';
+import 'package:grouping_project/space/presentation/view_models/user_data_provider.dart';
 import 'package:grouping_project/space/presentation/views/components/layout/activity_layout.dart';
 import 'package:intl/date_symbol_data_local.dart';
 import 'package:intl/intl.dart';
@@ -56,11 +56,22 @@ class _ActivityListFrameState extends State<ActivityListFrame> {
             child: SfCalendar(
               view: CalendarView.month,
               headerDateFormat: "y 年 MM 月",
-              // headerHeight: 25,
-              headerStyle: CalendarHeaderStyle(
-                  textStyle: TextStyle(fontSize: 15, color: widget.color)),
+              // headerStyle: CalendarHeaderStyle(
+              //     textStyle: TextStyle(fontSize: 15, color: widget.color)),
+              // timeSlotViewSettings: TimeSlotViewSettings(
+              //     timeInterval: Duration(hours: 1),
+              //     timeIntervalHeight: 50,
+              //     timeTextStyle: TextStyle(
+              //         fontSize: 15,
+              //         color: widget.color,
+              //         fontWeight: FontWeight.bold),
+              // ),
               monthViewSettings: const MonthViewSettings(
-                  numberOfWeeksInView: 4, dayFormat: 'EEE'),
+                numberOfWeeksInView: 4, 
+                dayFormat: 'EEE', 
+                // showAgenda: true,
+                
+              ),
               showDatePickerButton: true,
               showTodayButton: true,
               dataSource:
