+++ grouping_flutter/lib/space/presentation/views/frames/navigate_rail_frame.dart
@@ -4,7 +4,8 @@ import 'package:go_router/go_router.dart';
 import 'package:grouping_project/core/theme/color.dart';
 import 'package:grouping_project/space/presentation/view_models/create_workspace_view_model.dart';
 import 'package:grouping_project/space/presentation/view_models/join_workspace_view_model.dart';
-import 'package:grouping_project/space/presentation/view_models/user_page_view_model.dart';
+import 'package:grouping_project/space/presentation/view_models/space_view_model.dart';
+import 'package:grouping_project/space/presentation/view_models/user_data_provider.dart';
 import 'package:grouping_project/space/presentation/views/components/layout/dashboard_frame_layout.dart';
 import 'package:grouping_project/space/presentation/views/components/profile_avatar.dart';
 import 'package:grouping_project/space/presentation/views/components/dialogs/create_workspace_dialog.dart';
