+++ grouping_flutter/lib/space/presentation/views/frames/user_space/user_space_info_frame.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import 'package:grouping_project/space/presentation/view_models/user_page_view_model.dart';
+import 'package:grouping_project/space/presentation/view_models/user_data_provider.dart';
 import 'package:grouping_project/app/presentation/components/data_display/key_value_pair_widget.dart';
 import 'package:grouping_project/space/presentation/views/components/layout/dashboard_frame_layout.dart';
 import 'package:grouping_project/space/presentation/views/components/profile_avatar.dart';
