+++ grouping_flutter/lib/space/presentation/views/pages/user_page_view.dart
@@ -2,7 +2,8 @@ import 'package:flutter/material.dart';
 import 'package:grouping_project/space/presentation/view_models/create_workspace_view_model.dart';
 import 'package:grouping_project/space/presentation/view_models/join_workspace_view_model.dart';
 import 'package:grouping_project/space/presentation/view_models/setting_view_model.dart';
-import 'package:grouping_project/space/presentation/view_models/user_page_view_model.dart';
+import 'package:grouping_project/space/presentation/view_models/space_view_model.dart';
+import 'package:grouping_project/space/presentation/view_models/user_data_provider.dart';
 import 'package:grouping_project/space/presentation/views/components/layout/dashboard_frame_layout.dart';
 import 'package:grouping_project/space/presentation/views/components/layout/dashboard_layout.dart';
 import 'package:grouping_project/space/presentation/views/frames/navigate_rail_frame.dart';
