+++ grouping_flutter/lib/space/presentation/views/pages/workspace_page_view.dart
@@ -1,6 +1,7 @@
 import 'package:flutter/material.dart';
-import 'package:grouping_project/space/presentation/view_models/workspace_view_model.dart';
-import 'package:grouping_project/space/presentation/view_models/user_page_view_model.dart';
+import 'package:grouping_project/space/presentation/view_models/group_data_provider.dart';
+import 'package:grouping_project/space/presentation/view_models/space_view_model.dart';
+import 'package:grouping_project/space/presentation/view_models/user_data_provider.dart';
 import 'package:grouping_project/space/presentation/views/components/layout/dashboard_frame_layout.dart';
 import 'package:grouping_project/space/presentation/views/components/layout/dashboard_layout.dart';
 import 'package:grouping_project/space/presentation/views/frames/navigate_rail_frame.dart';
